docs(README) README 수정

+++ src/main/java/com/example/calendar_develop/service/ScheduleServiceImpl.java
@@ -1,16 +1,15 @@
 package com.example.calendar_develop.service;
 
+import com.example.calendar_develop.Exception.ScheduleNotFoundException;
 import com.example.calendar_develop.dto.ScheduleDto.CreateScheduleRequestDto;
 import com.example.calendar_develop.dto.ScheduleDto.CreateScheduleResponseDto;
-import com.example.calendar_develop.dto.ScheduleDto.PutScheduleRequestDto;
 import com.example.calendar_develop.dto.ScheduleDto.PutScheduleResponseDto;
 import com.example.calendar_develop.entity.Schedule;
 import com.example.calendar_develop.entity.User;
 import com.example.calendar_develop.repository.ScheduleRepository;
 import jakarta.persistence.EntityNotFoundException;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.dao.EmptyResultDataAccessException;
+
 
 @Service
 public class ScheduleServiceImpl implements ScheduleService {
@@ -44,10 +43,9 @@ public class ScheduleServiceImpl implements ScheduleService {
 
     @Override
     public void deleteById(Long id) {
-        try {
-            scheduleRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException Exception) {
-            System.out.println("일정을 찾을 수 없습니다.");
+        if(!scheduleRepository.existsById(id)) {
+            throw new ScheduleNotFoundException("일정을 찾을 수 없습니다.");
         }
+        scheduleRepository.deleteById(id);
     }
 }
